requirements.txt
streamlit==1.31.0
websockets==15.0.1
requests==2.32.5
protobuf==6.32.1
python-dotenv==1.1.1
pytz==2025.2
grpcio==1.75.1
grpcio-tools==1.75.1

